pyproject.toml
[project]
name = "rcm-stage1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "faiss-cpu>=1.11.0",
    "fastmcp>=2.9.2",
    "google-adk>=1.5.0",
    "google-genai>=1.24.0",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "sqlglot>=26.31.0",
]
